feat(slack): add api interaction helpers
