+++ packages/adapter-slack/src/api/boundary.test.ts
@@ -1,11 +1,21 @@
-import { readFile } from "node:fs/promises";
+import { readdir, readFile } from "node:fs/promises";
 import { describe, expect, it } from "vitest";
 
 describe("api import boundary", () => {
   it("does not import the full adapter or runtime packages", async () => {
-    const source = await readFile(new URL("./index.ts", import.meta.url), {
-      encoding: "utf8",
-    });
+    const directory = new URL(".", import.meta.url);
+    const files = await readdir(directory);
+    const source = (
+      await Promise.all(
+        files
+          .filter((file) => file.endsWith(".ts") && !file.endsWith(".test.ts"))
+          .map((file) =>
+            readFile(new URL(file, directory), {
+              encoding: "utf8",
+            })
+          )
+      )
+    ).join("\n");
 
     expect(source).not.toContain('from "chat"');
     expect(source).not.toContain("from '@chat-adapter/shared'");
